stm32: rename mod traits to macros.
