+++ embassy-stm32/src/macros.rs
